added update asset in controller

+++ assetmanagement/src/main/java/assetmanagement/repository/AddressRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import assetmanagement.model.Address;
 
-public interface AddressRepository extends CrudRepository<Address, Long> {
+public interface AddressRepository extends CrudRepository<Address, Integer> {
 
     @Override
     Iterable<Address> findAll();
