+++ assetmanagement/src/main/java/assetmanagement/repository/EmployeeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import assetmanagement.model.Employee;
 
-public interface EmployeeRepository extends CrudRepository<Employee, Long> {
+public interface EmployeeRepository extends CrudRepository<Employee, Integer> {
 }
