+++ assetmanagement/src/main/java/assetmanagement/repository/InterventionRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import assetmanagement.model.Intervention;
 
-public interface InterventionRepository extends CrudRepository<Intervention, Long> {
+public interface InterventionRepository extends CrudRepository<Intervention, Integer> {
 }
