+++ assetmanagement/src/main/java/assetmanagement/repository/SpecificationRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import assetmanagement.model.Specification;
 
-public interface SpecificationRepository extends CrudRepository<Specification, Long> {
+public interface SpecificationRepository extends CrudRepository<Specification, Integer> {
 }
